Add IdbRunner class to start and stop idb
