+++ maestro-ios/src/main/java/ios/idb/IdbRunner.kt
@@ -1,0 +1,9 @@
+package ios.idb
+
+import io.grpc.ManagedChannel
+
+interface IdbRunner {
+    fun stop(channel: ManagedChannel)
+
+    fun start(): ManagedChannel
+}
